Discord Update
Update to a discord server that we can maintain with auto moderation, anti-spam, scammer protections and such. Current server is in a deprecated state and has been for a couple of years. Now to the detriment to the project. Retire it as a legacy server and build a first class chat experience for ETC.